VoiceAgent/Voice-agent-backup/Voice-agent/requirements.txt
ffmpeg-python==0.2.0
audio-recorder-streamlit==0.0.8
gtts==2.5.1
pydub==0.25.1
pyttsx3==2.90
SpeechRecognition==3.10.1
streamlit==1.31.1
python-dotenv==1.0.1
flask==3.0.2
flask-cors==4.0.0
groq==0.20.0
langchain==0.1.13
langchain-core==0.1.40
langchain-groq==0.1.0
langchain-openai==0.0.8
langchain-pinecone==0.0.3
pinecone==3.0.2
httpx==0.26.0
chromadb==0.4.22
openai==1.12.0